Fix the bats test script
I had the wrong directories in the bats script, and this fixes that.
I also used [this trick](https://github.com/sstephenson/bats/issues/3#issuecomment-382664336)
to get the client to only run once instead of once per test, which
makes things go a lot faster.

+++ test/client_tests.bats
@@ -3,36 +3,48 @@
 # Run this from the top level directory, i.e.,
 # bats test/client_tests.bats
 
-cd src
+# We only want to run the client once at the start because
+# it takes quite a while, so the `setup` work only happens
+# one time.
+setup(){
+    if [ "$BATS_TEST_NUMBER" -eq 1 ]; then
+       cd src
 
-# Clean out any previously downloaded files.
-rm -f small.txt
-rm -f AsYouLikeIt.txt
-rm -f binary.jpg
+       # Clean out any previously downloaded files.
+       rm -f small.txt
+       rm -f AsYouLikeIt.txt
+       rm -f binary.jpg
 
-# (Re)compile the code
-rm -f segmentedfilesystem/*.class
-javac segmentedfilesystem/*.java
+       # (Re)compile the code
+       rm -f segmentedfilesystem/*.class
+       javac segmentedfilesystem/*.java
 
-# Run the client
-java segmentedfilesystem.Main
+       # Run the client
+       java segmentedfilesystem.Main csci-4409.morris.umn.edu
 
-cd ..
+       cd ..
+    fi
+}
+
 
 @test "Your client correctly assembled small.txt" {
-  run diff ../test/target-files/small.txt small.txt
+  # Uncomment this line if you want to see the result of
+  # the diff if this test is failing. Similar lines can
+  # help with the other tests.
+  # diff test/target-files/small.txt src/small.txt
+  run diff test/target-files/small.txt src/small.txt
 
   [ "$status" -eq 0 ]
 }
 
 @test "Your client correctly assembled AsYouLikeIt.txt" {
-  run diff ../test/target-files/AsYouLikeIt.txt AsYouLikeIt.txt
+  run diff test/target-files/AsYouLikeIt.txt src/AsYouLikeIt.txt
 
   [ "$status" -eq 0 ]
 }
 
 @test "Your client correctly assembled binary.jpg" {
-  run diff ../test/target-files/binary.jpg binary.jpg
+  run diff test/target-files/binary.jpg src/binary.jpg
 
   [ "$status" -eq 0 ]
 }
